requirements.txt
requests
python-docx
beautifulsoup4
python-dotenv
